personaplex-dgx-arm64/requirements.txt
# Core dependencies
# Note: PyTorch for ARM64 may need to be installed separately
# Check: https://pytorch.org/get-started/locally/ for ARM64 builds
# torch>=2.1.0  # Install separately for ARM64 compatibility
# torchaudio>=2.1.0  # Install separately for ARM64 compatibility
transformers>=4.35.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=4.24.0

# Server and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Audio processing
# torchaudio>=2.1.0  # Install separately (see above)
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
scipy>=1.11.0

# HuggingFace
huggingface-hub>=0.19.0
tokenizers>=0.15.0

# Utilities
pyyaml>=6.0.1
requests>=2.31.0
tqdm>=4.66.0
psutil>=5.9.0

# Optional optimizations
# tensorrt>=8.6.0  # Uncomment if TensorRT is available
# onnxruntime-gpu>=1.16.0  # Uncomment if ONNX is needed
